refactor: refactor logger dir and remove log file creation with its deps
