+++ .env.example
@@ -26,3 +26,6 @@ SMTP_HOST=''
 SMTP_PORT=''
 SMTP_USER=''
 SMTP_PASS=''
+
+# LOG LEVEL
+LOG_LEVEL="info"
