+++ package.json
@@ -47,7 +47,6 @@
     "tailwind-merge": "^3.5.0",
     "tw-animate-css": "^1.4.0",
     "winston": "^3.19.0",
-    "winston-daily-rotate-file": "^5.0.0",
     "zod": "^4.3.6"
   },
   "devDependencies": {
