+++ src/lib/logger.ts
@@ -1,40 +1,7 @@
 import { Prisma } from '@/generated/prisma/client';
-import fs from 'fs';
-import path from 'path';
-import { fileURLToPath } from 'url';
 import winston from 'winston';
-import 'winston-daily-rotate-file';
 
-const { combine, timestamp, json, colorize, errors } = winston.format;
-
-const __filename = fileURLToPath(import.meta.url);
-const logDir = path.dirname(__filename);
-
-if (!fs.existsSync(logDir)) {
-  fs.mkdirSync(logDir, { recursive: true });
-}
-
-// Define transports file for combined logs and error logs with daily rotation
-const combinedTransport = new winston.transports.DailyRotateFile({
-  filename: path.join(logDir, 'combined-%DATE%.log'),
-  datePattern: 'YYYY-MM-DD',
-  maxFiles: '7d',
-  format: combine(errors({ stack: true }), timestamp(), json()),
-});
-
-const errorTransport = new winston.transports.DailyRotateFile({
-  filename: path.join(logDir, 'error-%DATE%.log'),
-  datePattern: 'YYYY-MM-DD',
-  level: 'error',
-  maxFiles: '7d',
-  format: combine(errors({ stack: true }), timestamp(), json()),
-});
-
-export const logger = winston.createLogger({
-  level: process.env.LOG_LEVEL || 'info',
-  defaultMeta: { service: 'user-service' },
-  transports: [combinedTransport, errorTransport],
-});
+const { combine, timestamp, colorize, errors } = winston.format;
 
 const consoleFormat = combine(
   timestamp({ format: 'HH:mm:ss' }),
@@ -58,14 +25,15 @@ const consoleFormat = combine(
   }),
 );
 
-// terminal logging for development
-if (process.env.NODE_ENV !== 'production') {
-  logger.add(
+export const logger = winston.createLogger({
+  level: process.env.LOG_LEVEL || 'info',
+  defaultMeta: { service: 'user-service' },
+  transports: [
     new winston.transports.Console({
       format: consoleFormat,
     }),
-  );
-}
+  ],
+});
 
 const prisma_error_instance = [
   Prisma.PrismaClientKnownRequestError,
