+++ src/services/article.service.ts
@@ -1,5 +1,5 @@
 import { ApiError } from '@/lib/error';
-import { logger } from '@/logger/logger';
+import { logger } from '@/lib/logger';
 import { articleRepository } from '@/repositories/article.repository';
 import type {
   CreateArticleInput,
